CVPR/2025/StdGEN: Semantic-Decomposed 3D Character Generation from Single Images/requirements.txt
omegaconf
rm_anime_bg
diffusers==0.27.2
transformers==4.42.4
einops
huggingface_hub==0.25.0
opencv-python
accelerate
matplotlib
kornia
imageio
imageio-ffmpeg
xatlas
trimesh
rembg
onnxruntime
scikit-learn
pygltflib
pymeshlab
pytorch_lightning
git+https://github.com/NVlabs/nvdiffrast
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/Baijiong-Lin/LoRA-Torch
git+https://github.com/facebookresearch/segment-anything.git
